Version 0.2.0 - to_decimal now returns Option<u8> instead of u32 - added functions normalize_decimal and normalize_decimals, normalize_decimals_filtering

+++ packages/dec_from_char/benches/decimal_extended_benchmark.rs
@@ -33,7 +33,7 @@ fn benchmark_decimal_extended_string(c: &mut Criterion) {
             let _ = black_box(text)
                 .chars()
                 .filter_map(|c| c.to_decimal_utf8())
-                .collect::<Vec<u32>>();
+                .collect::<Vec<u8>>();
         })
     });
 }
